Split web localDb cards domain
