+++ apps/web/src/localDb/cards/cards/index.ts
@@ -2,13 +2,13 @@ import type {
   Card,
   QueryCardsInput,
   QueryCardsPage,
-} from "../../types";
+} from "../../../types";
 import {
   matchesCardFilter,
   matchesDeckFilterDefinition,
-} from "../../appData/domain";
-import { deriveDueAtBucketMillis, deriveDueAtMillis } from "../../appData/domain/dueAt";
-import { loadAllowedCardIdsForTags, putCardTagRecords, writeCardTagRecords } from "./cardTags";
+} from "../../../appData/domain";
+import { deriveDueAtBucketMillis, deriveDueAtMillis } from "../../../appData/domain/dueAt";
+import { loadAllowedCardIdsForTags, putCardTagRecords, writeCardTagRecords } from "../tags";
 import {
   closeDatabaseAfter,
   closeDatabaseAfterWrite,
@@ -16,8 +16,8 @@ import {
   getFromStore,
   runReadwrite,
   type StoredCard,
-} from "../core/database";
-import { encodeCursor, decodeCursor } from "../core/queryShared";
+} from "../../core/database";
+import { encodeCursor, decodeCursor } from "../../core/queryShared";
 
 type CardCursorIndexName =
   | "workspaceId_createdAt_cardId"
