+++ apps/web/src/localDb/cards/index.ts
@@ -1,4 +1,4 @@
-export * from "./cardTags";
 export * from "./cards";
 export * from "./decks";
+export * from "./tags";
 export * from "./workspace";
